フロント関係のファイルを再度置き換え
