+++ app/javascript/components/Article.vue
@@ -4,6 +4,13 @@
       <v-layout xs-12 class="top-info-container">
         <span class="user-name">@{{ article.user.name }}</span>
         <time-ago :refresh="60" :datetime="article.updated_at" locale="en" tooltip="top" long></time-ago>
+        <v-spacer></v-spacer>
+        <v-btn text fab small class="mr-5" @click="moveToEditArticlePage(article.id)">
+          <v-icon color="#3085DE">fas fa-pencil-alt</v-icon>
+        </v-btn>
+        <v-btn text fab small class="mr-2" @click="confirmDeleteArticle">
+          <v-icon color="#3085DE">fas fa-trash-alt</v-icon>
+        </v-btn>
       </v-layout>
       <v-layout>
         <h1 class="article-title">{{ article.title }}</h1>
@@ -20,6 +27,17 @@ import axios from "axios";
 import TimeAgo from 'vue2-timeago'
 import marked from "marked";
 import hljs from 'highlight.js';
+import Router from "../router/router";
+
+const headers = {
+  headers: {
+    Authorization: "Bearer",
+    "Access-Control-Allow-Origin": "*",
+    "access-token": localStorage.getItem("access-token"),
+    client: localStorage.getItem("client"),
+    uid: localStorage.getItem("uid")
+  }
+};
 
 export default {
   components: {
@@ -62,6 +80,10 @@ export default {
       return function(text) {
         return marked(text);
       };
+    },
+
+    editAble() {
+      return localStorage.getItem("uid") === this.article.user.email
     }
   },
 
@@ -77,6 +99,25 @@ export default {
           alert(e.response.statusText);
         });
     },
+
+    moveToEditArticlePage(id) {
+      Router.push(`/articles/${id}/edit`);
+    },
+
+    async confirmDeleteArticle() {
+      const result = confirm("この記事を削除してもよろしいですか？")
+      if (result) {
+        await axios
+          .delete(`/api/v1/articles/${this.article.id}`, headers)
+          .then(_response => {
+            Router.push("/")
+          })
+          .catch(e => {
+            // TODO: 適切な Error 表示
+            alert(e.response.statusText);
+          });
+      }
+    }
   }
 }
 </script>
