+++ app/javascript/components/MyPage.vue
