--- app/javascript/javascript/store/store.js
@@ -1,13 +1,0 @@
-import Vue from "vue";
-import Vuex from "vuex";
-
-Vue.use(Vuex);
-
-const store = new Vuex.Store({
-  state: {},
-  getters: {},
-  mutations: {},
-  actions: {},
-});
-
-export default store;
