+++ app/javascript/packs/hello_vue.js
@@ -15,6 +15,7 @@ import axios from "axios";
 import VueAxios from "vue-axios";
 import "vuetify/dist/vuetify.min.css";
 import "highlight.js/styles/monokai.css";
+import '@fortawesome/fontawesome-free/css/all.css'
 import Vuetify from "vuetify";
 Vue.use(Vuex);
 Vue.use(VueRouter);
