+++ config/routes.rb
@@ -6,6 +6,7 @@ Rails.application.routes.draw do
   get "sign_in", to: "home#index"
   get "articles/new", to: "home#index"
   get "articles/:id", to: "home#index"
+  get "mypage", to: "home#index"
 
   namespace :api do
     namespace :v1 do
